readstore_basic/requirements.txt
# readstore/django/requirements.py

Django==5.1
djangorestframework==3.15
djangorestframework-simplejwt==5.3
requests==2.32.3
gunicorn==23.0.0
pysam==0.22.1
pyyaml==6.0.2
streamlit==1.39.0
pydantic==2.9
pandas==2.2